+++ tool_packages/unique_web_search/CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), 
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.5.4] - 2025-11-12
+- Move pytest and pytest-asyncio to dev dependencies
+
+## [1.5.3] - 2025-11-10
+- Use SkipJsonSchema for mode under WebSearchMode config to prevent displaying an editable field
+
+## [1.5.2] - 2025-11-10
+- Separate the configuration of modes to prevent breaking the frontend
+
 ## [1.5.1] - 2025-11-10
 - Flag V2 mode and Advanced Query Refinement as Beta